Split out repo and page routes
