+++ app/constants/repos.json
@@ -1,0 +1,11 @@
+[
+  { "org": "onflow", "repo": "flow" },
+  { "org": "onflow", "repo": "cadence" },
+  { "org": "onflow", "repo": "flow-cli" },
+  { "org": "onflow", "repo": "flow-js-testing" },
+  { "org": "onflow", "repo": "flow-go-sdk" },
+  { "org": "onflow", "repo": "fcl-js" },
+  { "org": "onflow", "repo": "flow-emulator" },
+  { "org": "onflow", "repo": "flow-cadut" },
+  { "org": "onflow", "repo": "mock-developer-doc" }
+]
