--- app/routes/cadence/index.tsx
@@ -1,9 +1,0 @@
-import { LoaderFunction } from "@remix-run/server-runtime"
-
-export const loader: LoaderFunction = async () => {
-  return {}
-}
-
-export default function Page() {
-  return <div>Cadence Homepage</div>
-}
